requirements.txt
Django
django-allauth
django-crispy-forms
djangorestframework
Pillow
requests
python-telegram-bot